# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Api Gutter Runner",
        "description": "Implémentation de Swagger",
        "contact": {
            "email": "admin@app.gutterrunner.ca"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur d'API de démonstration"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Inscription"
                ],
                "summary": "Inscription d'un utilisateur",
                "description": "Inscription d'un utilisateur",
                "operationId": "userRegistration",
                "requestBody": {
                    "description": "Paramètres d'inscription d'un utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "payment_method",
                                    "plan"
                                ],
                                "properties": {
                                    "company_name": {
                                        "description": "Nom de la compagnie où l'utilisateur travaille",
                                        "type": "string",
                                        "example": "Alura"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "math@gmail.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe",
                                        "type": "string",
                                        "example": "Math2021!"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmation du mot de passe",
                                        "type": "string",
                                        "example": "Math2021!"
                                    },
                                    "payment_method": {
                                        "description": "Moyen de paiement",
                                        "type": "string",
                                        "example": "pm_1IpCtsL1HC1QKC81m2mqd8l3"
                                    },
                                    "plan": {
                                        "description": "Identifiant unique du plan d'abonnement",
                                        "type": "string",
                                        "example": "gutter_runner_monthly_forfait_gestionnaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Envoi d'email de vérification et abonnement réussi à un plan Stripe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Message de confirmation",
                                            "type": "string",
                                            "example": "Un courriel de vérification d'adresse email vous a été envoyé."
                                        },
                                        "subscription": {
                                            "description": "Détails de l'abonnement",
                                            "properties": {
                                                "name": {
                                                    "description": "Nom de l'abonnement",
                                                    "type": "string",
                                                    "example": "default"
                                                },
                                                "stripe_id": {
                                                    "description": "Identifiant unique de l'abonnement",
                                                    "type": "string",
                                                    "example": "sub_JS52aO9Xa7QhWw"
                                                },
                                                "stripe_status": {
                                                    "description": "Statut de l'abonnement",
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "stripe_plan": {
                                                    "description": "Identifiant du plan d'abonnement",
                                                    "type": "string",
                                                    "example": "gutter_runner_monthly_forfait_gestionnaire"
                                                },
                                                "quantity": {
                                                    "description": "Nombre d'unités de produits",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "trial_ends_at": {
                                                    "description": "Date de fin d'essai de l'abonnement",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "ends_at": {
                                                    "description": "Date de fin de l'abonnement",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "user_id": {
                                                    "description": "Identifiant de l'utilisateur abonné",
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "updated_at": {
                                                    "description": "Date de mis à jour de l'abonnement",
                                                    "type": "string",
                                                    "example": "2021-05-09T11:32:30.000000Z"
                                                },
                                                "created_at": {
                                                    "description": "Date de création de l'abonnement",
                                                    "type": "string",
                                                    "example": "2021-04-28T21:21:35.000000Z"
                                                },
                                                "id": {
                                                    "description": "Identifiant de l'abonnement dans la base de données",
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Paramètres vides ou invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Le champ adresse email doit être une adresse email valide."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Le champ de confirmation mot de passe ne correspond pas."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cartes invalides ou autres erreurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Your card has insufficient funds."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}